refactor: Move `World::get_reflect{_mut}` to new module `world::reflect`
Feedback from PR
